requirements.txt
# Core data manipulation
pandas>=2.0.0,<3.0.0
numpy>=1.24.0,<2.0.0
scipy>=1.10.0,<2.0.0

# Financial data
yfinance>=0.2.28,<0.3.0
pandas-datareader>=0.10.0,<1.0.0

# Statistical analysis
statsmodels>=0.14.0,<1.0.0

# Optimization
cvxpy>=1.3.0,<2.0.0
PyPortfolioOpt>=1.5.0,<2.0.0

# Visualization
matplotlib>=3.7.0,<4.0.0
plotly>=5.14.0,<6.0.0
seaborn>=0.12.0,<1.0.0

# Web dashboard
streamlit>=1.28.0,<2.0.0

# PDF generation
reportlab>=4.0.0,<5.0.0
weasyprint>=59.0,<63.0

# Configuration
python-dotenv>=1.0.0,<2.0.0

# HTTP requests with retry
requests>=2.31.0,<3.0.0
requests-cache>=1.1.0,<2.0.0

# Testing
pytest>=7.4.0,<9.0.0
pytest-cov>=4.1.0,<6.0.0
pytest-mock>=3.11.0,<4.0.0

# Data formats
pyarrow>=12.0.0,<18.0.0  # For parquet support
openpyxl>=3.1.0,<4.0.0   # For Excel support

# Utilities
tqdm>=4.65.0,<5.0.0      # Progress bars
colorama>=0.4.6,<1.0.0   # Colored terminal output

# Optional dependencies (uncomment if needed)
# scikit-learn>=1.3.0,<2.0.0  # For machine learning features
# jupyter>=1.0.0               # For notebooks
# black>=23.0.0                # Code formatting
# flake8>=6.0.0                # Linting
# mypy>=1.0.0                  # Type checking